Removing xyzCommon and reverting back here-xyz changes
