+++ .gitignore
@@ -3,4 +3,3 @@ node_modules/*
 bin/*.js
 .idea
 .vscode/launch.json
-bin/*.map
