+++ package.json
@@ -41,7 +41,6 @@
     "table": "^5.4.6",
     "tmp": "0.1.0",
     "user-settings": "0.2.0",
-    "weeknumber": "1.1.2",
     "xmldom": "0.3.0"
   },
   "devDependencies": {
