--- src/xyzCommon.ts
@@ -1,1169 +1,0 @@
-#!/usr/bin/env node
-
-/*
-  Copyright (C) 2018 - 2019 HERE Europe B.V.
-  SPDX-License-Identifier: MIT
-
-  Permission is hereby granted, free of charge, to any person obtaining
-  a copy of this software and associated documentation files (the
-  'Software'), to deal in the Software without restriction, including
-  without limitation the rights to use, copy, modify, merge, publish,
-  distribute, sublicense, and/or sell copies of the Software, and to
-  permit persons to whom the Software is furnished to do so, subject to
-  the following conditions:
-
-  The above copyright notice and this permission notice shall be
-  included in all copies or substantial portions of the Software.
-
-  THE SOFTWARE IS PROVIDED 'AS IS', WITHOUT WARRANTY OF ANY KIND,
-  EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-  MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-  IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-  CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-  TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-  SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-*/
-
-import * as common from "./common";
-import * as transform from "./transformutil";
-import {requestAsync} from "./requestAsync";
-import {ApiError} from "./api-error";
-import * as zlib from "zlib";
-import * as inquirer from "inquirer";
-const path = require('path');
-const weeknumber = require('weeknumber');
-import * as summary from "./summary";
-import * as fs from "fs";
-import * as glob from "glob";
-import * as moment from "moment";
-
-const gsv = require("geojson-validation");
-let cq = require("block-queue");
-
-export const bboxDirections = ["west", "south", "east", "north"];
-export let choiceList: { name: string, value: string }[] = [];
-export const commandHistoryCount = 3;
-
-const filesToUpload = [
-    {
-        type: "checkbox",
-        name: "selectedFiles",
-        message: "Select the files to be uploaded",
-        choices: choiceList
-    }
-];
-
-export const questionConfirm = [
-    {
-        type: 'input',
-        name: 'confirmed',
-        message: 'Enter (Y)es to continue or (N)o to cancel'
-    }
-];
-
-export async function createSpace(options: any) {
-
-    if (options) {
-        if (!options.title) {
-            options.title = "a new Data Hub space created from commandline";
-        }
-        if (!options.message) {
-            options.message = "a new Data Hub space created from commandline";
-        }
-    }
-    let gp: any = getGeoSpaceProfiles(options.title, options.message, options.client);
-
-    if (options.schema) {
-
-        await common.verifyProLicense();
-
-        if (options.schema == true) {
-            console.log("Please add local filepath / http link for your schema definition")
-            process.exit(1);
-        } else {
-            let schemaDef: string = "";
-            if (options.schema.indexOf("http") == 0) {
-                schemaDef = options.schema;
-            } else {
-                schemaDef = await transform.read(options.schema, false);
-            }
-
-            schemaDef = schemaDef.replace(/\r?\n|\r/g, " ");
-
-            let processors = getSchemaProcessorProfile(schemaDef);
-
-            gp['processors'] = processors;
-        }
-    }
-
-
-    const response = await execute("/hub/spaces?clientId=cli", "POST", "application/json", gp, options.token);
-    console.log("Data Hub space '" + response.body.id + "' created successfully");
-    return response.body;
-}
-
-/**
- *
- * @param apiError error object
- * @param isIdSpaceId set this boolean flag as true if you want to give space specific message in console for 404
- */
-export function handleError(apiError: ApiError, isIdSpaceId: boolean = false) {
-    if (apiError.statusCode) {
-        if (apiError.statusCode == 401) {
-            console.log("Operation FAILED : Unauthorized, if the problem persists, please reconfigure account with `here configure` command");
-        } else if (apiError.statusCode == 403) {
-            console.log("Operation FAILED : Insufficient rights to perform action");
-        } else if (apiError.statusCode == 404) {
-            if (isIdSpaceId) {
-                console.log("Operation FAILED: Space does not exist");
-            } else {
-                console.log("Operation FAILED : Resource not found.");
-            }
-        } else {
-            console.log("OPERATION FAILED : " + apiError.message);
-        }
-    } else {
-        if (apiError.message && apiError.message.indexOf("Insufficient rights.") != -1) {
-            console.log("Operation FAILED - Insufficient rights to perform action");
-        } else {
-            console.log("OPERATION FAILED - " + apiError.message);
-        }
-    }
-}
-
-export async function execute(uri: string, method: string, contentType: string, data: any, token: string | null = null, gzip: boolean = false, setAuthorization: boolean = true) {
-    if (!token) {
-        token = await common.verify();
-    }
-    return await execInternal(uri, method, contentType, data, token, gzip, setAuthorization);
-}
-
-export async function execInternal(
-    uri: string,
-    method: string,
-    contentType: string,
-    data: any,
-    token: string,
-    gzip: boolean,
-    setAuthorization: boolean
-) {
-    if (gzip) {
-        return await execInternalGzip(
-            uri,
-            method,
-            contentType,
-            data,
-            token
-        );
-    }
-    if (!uri.startsWith("http")) {
-        uri = common.xyzRoot() + uri;
-    }
-    const responseType = contentType.indexOf('json') !== -1 ? 'json' : 'text';
-    const reqJson = {
-        url: uri,
-        method: method,
-        headers: {
-            Authorization: "Bearer " + token,
-            "Content-Type": contentType,
-            "App-Name": "HereCLI"
-        },
-        json: method === "GET" ? undefined : data,
-        allowGetBody: true,
-        responseType: responseType
-    };
-
-    //Remove Auth params if not required, Used to get public response from URL
-    if (setAuthorization == false) {
-        delete reqJson.headers.Authorization;
-    }
-
-    const response = await requestAsync(reqJson);
-    if (response.statusCode < 200 || response.statusCode > 210) {
-        let message = (response.body && response.body.constructor != String) ? JSON.stringify(response.body) : response.body;
-        //throw new Error("Invalid response - " + message);
-        throw new ApiError(response.statusCode, message);
-    }
-    return response;
-}
-
-export async function execInternalGzip(
-    uri: string,
-    method: string,
-    contentType: string,
-    data: any,
-    token: string,
-    retry: number = 3
-) {
-    const zippedData = await gzip(data);
-    if (!uri.startsWith("http")) {
-        uri = common.xyzRoot() + uri;
-    }
-    const responseType = contentType.indexOf('json') !== -1 ? 'json' : 'text';
-    const reqJson = {
-        url: uri,
-        method: method,
-        headers: {
-            Authorization: "Bearer " + token,
-            "Content-Type": contentType,
-            "Content-Encoding": "gzip",
-            "Accept-Encoding": "gzip"
-        },
-        decompress: true,
-        body: method === "GET" ? undefined : zippedData,
-        allowGetBody: true,
-        responseType: responseType
-    };
-
-    let response = await requestAsync(reqJson);
-    if (response.statusCode < 200 || response.statusCode > 210) {
-        if (response.statusCode >= 500 && retry > 0) {
-            await new Promise(done => setTimeout(done, 1000));
-            response = await execInternalGzip(uri, method, contentType, data, token, --retry);
-        } else {
-            //   throw new Error("Invalid response :" + response.statusCode);
-            throw new ApiError(response.statusCode, response.body);
-        }
-    }
-    return response;
-}
-
-export function gzip(data: zlib.InputType): Promise<Buffer> {
-    return new Promise<Buffer>((resolve, reject) =>
-        zlib.gzip(data, (error, result) => {
-            if (error)
-                reject(error)
-            else
-                resolve(result);
-        })
-    );
-}
-
-export function getGeoSpaceProfiles(title: string, description: string, client: any) {
-    return {
-        title,
-        description,
-        client,
-        "enableUUID": true
-    };
-}
-
-export function getSchemaProcessorProfile(schema: string) {
-    return {
-        "schema-validator" : [{
-            "eventTypes": ["ModifyFeaturesEvent.request", "ModifySpaceEvent.request"],
-            "params": {
-                "schema": schema
-            },
-            "order": 0
-        }]
-    }
-}
-
-export function getSpaceDataFromXyz(id: string, options: any) {
-    return new Promise<any>(function (resolve, reject) {
-        let cType = "application/json";
-        if (!options.limit) {
-            options.limit = 5000;
-        }
-        const getUrI = function (offset: string) {
-            let uri = "/hub/spaces/" + id;
-            let spFunction;
-            if (options.bbox) {
-                spFunction = "bbox";
-                options.limit = 100000;//Max limit of records space api supports
-            } else {
-                spFunction = "iterate";
-            }
-            if (options.limit) {
-                uri = uri + "/" + spFunction + "?limit=" + options.limit + "&clientId=cli";
-                if (options.bbox) {
-                    var bboxarray = options.bbox.split(",");
-                    if (bboxarray.length !== 4) {
-                        console.error(`\nboundingbox input size is not proper - "${options.bbox}"`);
-                        process.exit(1);
-                    }
-                    bboxarray.forEach(function (item: string, i: number) {
-                        if (item && item != "") {
-                            let number = parseFloat(item.toLowerCase());
-                            if (isNaN(number)) {
-                                console.error(`\nLoading space data using bounding box failed - "${item}" is not a valid number`);
-                                process.exit(1);
-                            }
-                            uri = uri + "&" + bboxDirections[i] + "=" + number;
-                        }
-                    });
-                }
-                if (offset) {
-                    uri = uri + "&handle=" + offset;
-                }
-                if (options.tags) {
-                    uri = uri + "&tags=" + options.tags;
-                }
-            }
-            return uri;
-        };
-        if (!options.totalRecords) {
-            options.totalRecords = 500000;
-        }
-        let recordLength = 0;
-        let features = new Array();
-        let jsonOut;
-        (async () => {
-
-            try {
-                let cHandle = options.handle ? options.handle : 0;
-                if (cHandle === 0 && !options.ignoreLogs) {
-                    process.stdout.write("Operation may take a while. Please wait...");
-                }
-                do {
-                    if(!options.ignoreLogs){
-                        process.stdout.write(".");
-                    }
-                    let response = await execute(
-                        getUrI(String(cHandle)),
-                        "GET",
-                        cType,
-                        "",
-                        options.token,
-                        true
-                    );
-                    jsonOut = response.body;
-                    if (jsonOut.constructor !== {}.constructor) {
-                        jsonOut = JSON.parse(jsonOut);
-                    }
-                    cHandle = jsonOut.handle;
-                    if (jsonOut.features) {
-                        recordLength += jsonOut.features.length;
-                        features = features.concat(jsonOut.features);
-                    } else {
-                        cHandle = -1;
-                    }
-                    if (options.currentHandleOnly) {
-                        cHandle = -1;
-                        break;
-                    }
-                } while (cHandle >= 0 && recordLength < options.totalRecords);
-                if (!options.currentHandleOnly) {
-                    process.stdout.write("\n");
-                }
-                jsonOut.features = features;
-                resolve(jsonOut);
-            } catch (error) {
-                console.error(`\ngetting data from Data Hub space failed: ${JSON.stringify(error)}`);
-                reject(error);
-            }
-        })();
-    });
-}
-
-export async function getSpaceMetaData(id: string, token: string | null = null) {
-    const uri = "/hub/spaces/" + id + "?clientId=cli";
-    const cType = "application/json";
-    const response = await execute(uri, "GET", cType, "", token);
-    return response.body;
-}
-
-export async function getStatisticsData(spaceId: string, token: string | null = null) {
-    const response = await execute(
-        "/hub/spaces/" + spaceId + "/statistics",
-        "GET",
-        "application/json",
-        null,
-        token,
-        true
-    );
-    return response.body;
-}
-
-export async function uploadToXyzSpace(id: string, options: any) {
-    //(async () => {
-    let tags = "";
-    if (options.tags) {
-        tags = options.tags;
-    }
-
-    let printErrors = false;
-    if (options.errors) {
-        printErrors = true;
-    }
-
-    //Default chunk size set as 200
-    if (!options.chunk) {
-        options.chunk = 200;
-    }
-
-    if (options.unique && options.override) {
-        console.log(
-            "conflicting options -- you must use either unique or override. Refer to 'here xyz upload -h' for help"
-        );
-        process.exit(1);
-    } else if (!options.override) {
-        options.unique = true;
-    }
-
-    if (options.assign && options.stream) {
-        console.log(
-            "conflicting options - you cannot choose assign mode while selecting streaming option"
-        );
-        process.exit(1);
-    }
-
-    let files: string[] = [''];//Initialising as blank string, so that if options.file is not given loop will execute atleast once and else condition will be executed
-    if(options.batch){
-        files = [];
-        if(options.batch != true && options.batch.toLowerCase().indexOf(".") == -1){
-            options.batch = "*."+options.batch;
-        }
-        let directories = options.file.split(',');
-        for(let directory of directories) {
-            if(!(fs.existsSync(directory) && fs.lstatSync(directory).isDirectory())){
-                console.log("--batch option requires directory path in --file option");
-                process.exit(1);
-            }
-            if(options.batch == true){
-                const allFiles = fs.readdirSync(directory, { withFileTypes: true })
-                    .filter(dirent => dirent.isFile())
-                    .map(dirent => dirent.name);
-                allFiles.forEach(function (item: any) {
-                    choiceList.push({'name': item, 'value': path.join(directory,item)});
-                });
-            } else {
-                files = files.concat(glob.sync(path.join(directory,options.batch)));
-                if(options.batch == 'shp' || options.batch == '*.shp'){
-                    const allDirectories = fs.readdirSync(directory, { withFileTypes: true })
-                        .filter(dirent => dirent.isDirectory())
-                        .map(dirent => dirent.name);
-                    for(let subDirectory of allDirectories) {
-                        files = files.concat(glob.sync(path.join(directory,subDirectory,options.batch)));
-                    }
-                }
-            }
-        }
-        if(options.batch == true){
-            let answers: any = await inquirer.prompt(filesToUpload);
-            files = answers.selectedFiles;
-        }
-        if(files.length == 0){
-            console.log("No files found of the specified format in the directory");
-            process.exit(1);
-        }
-    } else if(options.file){
-        files = options.file.split(',');
-    }
-
-    for(let file of files) {
-        options.file = file;
-        console.log("uploading file - " + file);
-        let startTime = new Date();
-        if(!options.stream && options.file && !(options.file.toLowerCase().indexOf(".shp") != -1 || options.file.toLowerCase().indexOf(".gpx") != -1)){
-            console.log("you can stream your uploads of CSV, GeoJSON and GeoJSONL files using the -s option. This will allow you to upload very large files, and will dramatically reduce the upload time for files of any size.");
-        }
-
-        if (options.file) {
-            const fs = require("fs");
-            if (options.file.toLowerCase().indexOf(".geojsonl") != -1) {
-                if (!options.stream) {
-                    const result: any = await transform.readLineFromFile(options.file, 100);
-                    await uploadData(id, options, tags, { type: "FeatureCollection", features: collate(result) }, true, options.ptag, options.file, options.id, printErrors);
-                } else {
-                    let queue = streamingQueue();
-                    await transform.readLineAsChunks(options.file, options.chunk ? options.chunk : 1000, options, function (result: any) {
-                        return new Promise((res, rej) => {
-                            (async () => {
-                                if (result.length > 0) {
-                                    await queue.send({ id: id, options: options, tags: tags, fc: { type: "FeatureCollection", features: collate(result) }, retryCount: 3 });
-                                }
-                                res(queue);
-                            })();
-                        });
-                    });
-                    while (queue.chunksize != 0) {
-                        await new Promise(done => setTimeout(done, 1000));
-                    }
-                }
-            } else if (options.file.toLowerCase().indexOf(".shp") != -1) {
-                let result = await transform.readShapeFile(
-                    options.file,
-                );
-                await uploadData(
-                    id,
-                    options,
-                    tags,
-                    result,
-                    true,
-                    options.ptag,
-                    options.file,
-                    options.id
-                );
-            } else if (options.file.toLowerCase().indexOf(".csv") != -1 || options.file.toLowerCase().indexOf(".txt") != -1) {
-                if (!options.stream) {
-                    let result = await transform.read(
-                        options.file,
-                        true,
-                        { headers: true, delimiter: options.delimiter, quote: options.quote }
-                    );
-                    const object = {
-                        features: await transform.transform(
-                            result,
-                            options
-                        ),
-                        type: "FeatureCollection"
-                    };
-                    await uploadData(
-                        id,
-                        options,
-                        tags,
-                        object,
-                        true,
-                        options.ptag,
-                        options.file,
-                        options.id
-                    );
-                } else {
-                    let queue = streamingQueue();
-                    await transform.readCSVAsChunks(options.file, options.chunk ? options.chunk : 1000, options, function (result: any) {
-                        return new Promise((res, rej) => {
-                            (async () => {
-                                if (result.length > 0) {
-                                    const fc = {
-                                        features: await transform.transform(
-                                            result,
-                                            options
-                                        ),
-                                        type: "FeatureCollection"
-                                    };
-                                    await queue.send({ id: id, options: options, tags: tags, fc: fc, retryCount: 3 });
-                                }
-                                res(queue);
-                            })();
-                        });
-
-                    });
-                    while (queue.chunksize != 0) {
-                        await new Promise(done => setTimeout(done, 1000));
-                    }
-                }
-            } else if (options.file.indexOf(".gpx") != -1) {
-                let result = await transform.read(
-                    options.file,
-                    false,
-                    {}
-                );
-                const object = {
-                    features: await transform.transformGpx(
-                        result,
-                        options
-                    ),
-                    type: "FeatureCollection"
-                };
-                await uploadData(
-                    id,
-                    options,
-                    tags,
-                    object,
-                    true,
-                    options.ptag,
-                    options.file,
-                    options.id
-                );
-
-            } else {
-                if (!options.stream) {
-                    let result = await transform.read(
-                        options.file,
-                        false
-                    );
-                    let object = JSON.parse(result);
-                    if(!(object.features && object.features.length > 0 && object.features[0].type == 'Feature')){
-                        object = {
-                            features: await transform.transform(
-                                object,
-                                options
-                            ),
-                            type: "FeatureCollection"
-                        };
-                    }
-                    await uploadData(
-                        id,
-                        options,
-                        tags,
-                        object,
-                        true,
-                        options.ptag,
-                        options.file,
-                        options.id
-                    );
-                } else {
-                    let queue = streamingQueue();
-                    let c = 0;
-                    await transform.readGeoJsonAsChunks(options.file, options.chunk ? options.chunk : 1000, options, async function (result: any) {
-                        if (result.length > 0) {
-                            const fc = {
-                                features: result,
-                                type: "FeatureCollection"
-                            };
-                            await queue.send({ id: id, options: options, tags: tags, fc: fc, retryCount: 3 });
-                        }
-                        return queue;
-                    });
-                    while (queue.chunksize != 0) {
-                        await new Promise(done => setTimeout(done, 1000));
-                    }
-                }
-            }
-        } else {
-            const getStdin = require("get-stdin");
-            await getStdin().then((str: string) => {
-                try {
-                    const obj = JSON.parse(str);
-                    uploadData(
-                        id,
-                        options,
-                        tags,
-                        obj,
-                        false,
-                        options.ptag,
-                        null,
-                        options.id
-                    );
-                } catch (e) {
-                    console.log(
-                        "Empty or invalid input to upload. Refer to 'here xyz upload -h' for help"
-                    );
-                    process.exit(1);
-                }
-            });
-        }
-
-        let totalTime = ((new Date().getTime() - startTime.getTime()) / 1000);
-        console.log(options.totalCount + " features uploaded to Data Hub space '" + id + "' in " + totalTime + " seconds, at the rate of " + Math.round(options.totalCount / totalTime) + " features per second");
-    }
-    await updateCommandMetadata(id, options, false, null);
-    console.log("upload completed successfully");
-    //})();
-}
-
-
-export async function updateCommandMetadata(id: string, options: any, isClear: boolean = false, favCommand: string | null = null){
-    let history: Array<any> = [];
-    let data: any = {};
-    if(favCommand){
-        data = {
-            client: {
-                'favouriteCommand': favCommand
-            }
-        }
-    } else {
-        let spaceData = await getSpaceMetaData(id, options.token);
-        if(spaceData.client && spaceData.client.history){
-            history = spaceData.client.history;
-        }
-        let command = {
-            "command" : `here xyz upload ${id} ` + process.argv.slice(4).map(x => x.includes(' ') ? "'" + x.trim() + "'": x.trim()).join(" "),
-            "timestamp": moment().toISOString(true)
-        }
-        history = [command].concat(history);
-        data = {
-            client: {
-                'history' : isClear ? [] : history.slice(0, commandHistoryCount)
-            }
-        }
-    }
-    const uri = "/hub/spaces/" + id + "?clientId=cli";
-    const cType = "application/json";
-    const response = await execute(uri, "PATCH", cType, data);
-    return response.body;
-}
-
-export function streamingQueue() {
-    let queue = cq(10, function (task: any, done: Function) {
-        uploadData(task.id, task.options, task.tags, task.fc,
-            true, task.options.ptag, task.options.file, task.options.id)
-            .then((result: any) => {
-                queue.uploadCount += result.success;
-                queue.failedCount += result.failed;
-                process.stdout.write("\ruploaded feature count :" + queue.uploadCount + ", failed feature count :" + queue.failedCount);
-                queue.chunksize--;
-                done();
-            }).catch((err: any) => {
-            queue.failedCount += task.fc.features.length;
-            process.stdout.write("\ruploaded feature count :" + queue.uploadCount + ", failed feature count :" + queue.failedCount);
-            queue.chunksize--;
-            done();
-        });
-    });
-    queue.uploadCount = 0;
-    queue.chunksize = 0;
-    queue.failedCount = 0;
-    queue.send = async function (obj: any) {
-        while (this.chunksize > 25) {
-            await new Promise(done => setTimeout(done, 1000));
-        }
-        this.push(obj);
-        this.chunksize++;
-    }
-    queue.shutdown = async () => {
-        queue.shutdown = true;
-        while (queue.chunksize != 0) {
-            await new Promise(done => setTimeout(done, 1000));
-        }
-        return true;
-    }
-    return queue;
-}
-
-export function collate(result: Array<any>) {
-    return result.reduce((features: any, feature: any) => {
-        if (feature.type === "Feature") {
-            features.push(feature);
-        } else if (feature.type === "FeatureCollection") {
-            features = features.concat(feature.features);
-        } else {
-            console.log("Unknown type" + feature.type);
-        }
-        return features
-    }, []);
-}
-
-export async function iterateChunks(chunks: any, url: string, index: number, chunkSize: number, token: string, upresult: any, printFailed: boolean): Promise<any> {
-    const item = chunks.shift();
-    const fc = { type: "FeatureCollection", features: item };
-    const response = await execute(
-        url,
-        "POST",
-        "application/geo+json",
-        JSON.stringify(fc, (key, value) => {
-            if (typeof value === 'string') {
-                return value.replace(/\0/g, '');
-            }
-            return value;
-        }),
-        token,
-        true
-    );
-
-    if (response.statusCode >= 200 && response.statusCode < 210) {
-        let res = response.body;
-        if (res.features)
-            upresult.success = upresult.success + res.features.length;
-        if (res.failed) {
-            upresult.failed = upresult.failed + res.failed.length;
-            //upresult.entries = upresult.entries.concat(res.failed);
-
-
-            for (let n = 0; n < res.failed.length; n++) {
-                const failedentry = res.failed[n];
-                if (printFailed) {
-                    console.log("Failed to upload : " + JSON.stringify({ feature: fc.features[failedentry.position], reason: failedentry.message }));
-                }
-            }
-        }
-    }
-    index++;
-    process.stdout.write("\ruploaded " + ((index / chunkSize) * 100).toFixed(2) + "%");
-    if (index == chunkSize) {
-        return upresult;
-    }
-    return await iterateChunks(chunks, url, index, chunkSize, token, upresult, printFailed);
-}
-
-export function chunkify(data: any[], chunksize: number) {
-    let chunks: any[] = [];
-    for (const k in data) {
-        const item = data[k];
-        if (!chunks.length || chunks[chunks.length - 1].length == chunksize)
-            chunks.push([]);
-        chunks[chunks.length - 1].push(item);
-    }
-    return chunks;
-}
-
-export async function uploadDataToSpaceWithTags(
-    id: string,
-    options: any,
-    tags: any,
-    object: any,
-    isFile: boolean,
-    tagProperties: any,
-    fileName: string | null,
-    uid: string,
-    upresult: any,
-    printFailed: boolean
-) {
-    return new Promise((resolve, reject) => {
-        gsv.valid(object, async function (valid: boolean, errs: any) {
-            if (!valid) {
-                console.log(errs);
-                reject(errs);
-                return;
-            }
-            const featureOut = await mergeAllTags(
-                object.features,
-                tags,
-                tagProperties,
-                fileName,
-                uid,
-                options
-            );
-
-            try {
-                if (options.stream) {
-                    upresult = await iterateChunks([featureOut], "/hub/spaces/" + id + "/features" + "?clientId=cli", 0, 1, options.token, upresult, printFailed);
-                } else {
-                    const chunks = options.chunk
-                        ? chunkify(featureOut, parseInt(options.chunk))
-                        : [featureOut];
-                    upresult = await iterateChunks(chunks, "/hub/spaces/" + id + "/features" + "?clientId=cli", 0, chunks.length, options.token, upresult, printFailed);
-                    process.stdout.write("\n");
-                    // let tq =  taskQueue(8,chunks.length);
-                    // chunks.forEach(chunk=>{
-                    //     tq.send({chunk:chunk,url:"/hub/spaces/" + id + "/features"});
-                    // });
-                    // await tq.shutdown();
-                }
-            } catch (e) {
-                reject(e);
-                return;
-            }
-
-            if (!options.stream) {
-                if (isFile)
-                    console.log(
-                        "'" +
-                        options.file +
-                        "' uploaded to XYZ space '" +
-                        id +
-                        "'"
-                    );
-                else
-                    console.log(
-                        "data upload to Data Hub space '" + id + "' completed"
-                    );
-
-                if (upresult.failed > 0) {
-                    console.log("all the features could not be uploaded successfully, to print rejected features, run command with -e")
-                    console.log("=============== Upload Summary ============= ");
-                    upresult.total = featureOut.length;
-                    console.table(upresult);
-                } else {
-                    summary.summarize(featureOut, id, true);
-                }
-                options.totalCount = featureOut.length;
-
-            }
-            resolve(upresult);
-        });
-    });
-}
-
-
-export async function mergeAllTags(
-    features: any,
-    tags: string,
-    tagProperties: any,
-    fileName: string | null,
-    idStr: string,
-    options: any,
-) {
-    let inputTags: Array<string> = [];
-    tags.split(",").forEach(function (item) {
-        if (item && item != "") inputTags.push(item.toLowerCase());
-    });
-    const tps = tagProperties ? tagProperties.split(",") : null;
-    let checkId = false;
-    const featureMap: Array<string> = [];
-    const duplicates = new Array();
-    features.forEach(function (item: any) {
-        let finalTags = inputTags.slice();
-        let origId = null;
-        //Generate id only if doesnt exist
-        if (!item.id && idStr) {
-            const fId = createUniqueId(idStr, item);
-            if (fId && fId != "") {
-                item.id = fId;
-            }
-        } else {
-            if (options.unique) {
-                checkId = true;
-                origId = item.id;
-                item.id = undefined;
-                const id = common.md5Sum(JSON.stringify(item));
-                item.id = id;
-                if (featureMap[item.id]) {
-                    const dupe = {
-                        id: origId,
-                        geometry: JSON.stringify(item.geometry),
-                        properties: JSON.stringify(item.properties)
-                    };
-                    duplicates.push(dupe);
-                }
-            }
-        }
-        if (options.unique) {
-            if (!featureMap[item.id]) {
-                featureMap[item.id] = item;
-            }
-        }
-        if (!item.properties) {
-            item.properties = {};
-        }
-        let metaProps = item.properties["@ns:com:here:xyz"];
-        if (!metaProps) {
-            metaProps = {};
-        }
-        if (metaProps && metaProps.tags) {
-            finalTags = finalTags.concat(metaProps.tags);
-        }
-        if (tps) {
-            tps.forEach(function (tp: any) {
-                if (item.properties[tp]) {
-                    if (Array.isArray(item.properties[tp])) {
-                        for (let i in item.properties[tp]) {
-                            addTagsToList(item.properties[tp][i], tp, finalTags);
-                        }
-                    } else {
-                        addTagsToList(item.properties[tp], tp, finalTags);
-                    }
-                }
-            });
-        }
-
-        if(options.date){
-            try{
-                options.date.split(",").forEach((element: any) => {
-                    const value = item.properties[element];
-                    if(value){
-                        let dateValue: Date;
-                        if(!isNaN(Number(value)) && !isNaN(parseFloat(value)) && isFinite(parseFloat(value))){
-                            dateValue = new Date(parseFloat(value.toString()));
-                        } else {
-                            dateValue = new Date(value);
-                        }
-                        item.properties['xyz_timestamp_'+element] = dateValue.getTime();
-                        item.properties['xyz_iso8601_'+element] = dateValue.toISOString();
-                        if(options.datetag){
-                            let allTags: boolean = false;
-                            if(options.datetag == true || options.datetag == undefined){
-                                allTags = true;
-                            }
-                            let inputTagsList = [];
-                            if(!allTags){
-                                inputTagsList = options.datetag.split(',');
-                            }
-                            if(allTags || inputTagsList.includes('year')){
-                                addTagsToList(dateValue.getUTCFullYear().toString(), 'date_'+element+'_year', finalTags);
-                            }
-                            if(allTags || inputTagsList.includes('month')){
-                                addTagsToList(dateValue.toLocaleString('UTC', { month: 'long' }), 'date_'+element+'_month', finalTags);
-                            }
-                            if(allTags || inputTagsList.includes('year_month')){
-                                addTagsToList(dateValue.getUTCFullYear().toString() + '-' + ("0" + (dateValue.getUTCMonth()+　1)).slice(-2).toString(), 'date_'+element+'_year_month', finalTags);
-                            }
-                            if(allTags || inputTagsList.includes('week')){
-                                addTagsToList(("0" + (weeknumber.weekNumber(dateValue)+　1)).slice(-2), 'date_'+element+'_week', finalTags);
-                            }
-                            if(allTags || inputTagsList.includes('year_week')){
-                                addTagsToList(dateValue.getUTCFullYear().toString() + '-' + ("0" + (weeknumber.weekNumber(dateValue)+　1)).slice(-2), 'date_'+element+'_year_week', finalTags);
-                            }
-                            if(allTags || inputTagsList.includes('weekday')){
-                                addTagsToList(dateValue.toLocaleString('UTC', { weekday: 'long' }), 'date_'+element+'_weekday', finalTags);
-                            }
-                        }
-                    }
-                });
-            } catch(e){
-                console.log("Invalid time format - " + e.message);
-                process.exit(1);
-            }
-        }
-        const nameTag = fileName ? getFileName(fileName) : null;
-        if (nameTag) {
-            finalTags.push(nameTag);
-        }
-        if (origId) {
-            metaProps.originalFeatureId = origId;
-        }
-        metaProps.tags = uniqArray(finalTags);
-        item.properties["@ns:com:here:xyz"] = metaProps;
-    });
-
-    if (options.unique && duplicates.length > 0) {
-        const featuresOut = new Array();
-        for (const k in featureMap) {
-            featuresOut.push(featureMap[k]);
-        }
-        console.log(
-            "***************************************************************"
-        );
-        console.log(
-            "We detected duplicate features in this chunk and only the first was uploaded. Features that had duplicates:\n"
-        );
-        common.drawTable(duplicates, ["id", "geometry", "properties"]); // TODO: suppress geometry of lines,polygons
-        console.log(
-            "uploading " +
-            featuresOut.length +
-            " out of " +
-            features.length +
-            " records"
-        );
-        console.log(
-            "***************************************************************\n"
-        );
-        return featuresOut;
-    } else {
-        return features;
-    }
-}
-
-export function uniqArray<T>(a: Array<T>) {
-    return Array.from(new Set(a));
-}
-
-export function getFileName(fileName: string) {
-    try {
-        const path = require("path");
-        let bName = path.basename(fileName);
-        if (bName.indexOf(".") != -1) {
-            bName = bName.substring(0, bName.lastIndexOf("."));
-        }
-        return bName;
-    } catch (e) {
-        return null;
-    }
-}
-
-export function addTagsToList(value: string, tp: string, finalTags: string[]) {
-    value = value.toString().toLowerCase();
-    value = value.replace(/\s+/g, "_");
-    value = value.replace(/,+/g, "_");
-    value = value.replace(/&+/g, "_and_");
-    value = value.replace(/\++/g, "_plus_");
-    value = value.replace(/#+/g, "_num_");
-    tp = tp.replace(/\s+/g, "_");
-    //finalTags.push(value); // should we add tags with no @ an option?
-    finalTags.push(tp + "@" + value);
-    return finalTags;
-}
-
-export function createUniqueId(idStr: string, item: any) {
-    const ids = idStr.split(",");
-    const vals = new Array();
-    ids.forEach(function (id) {
-        const v = item.properties ? item.properties[id] : null;
-        if (v) {
-            vals.push(v);
-        }
-    });
-    const idFinal = vals.join("-");
-    return idFinal;
-}
-
-export const questions = [
-    {
-        type: "checkbox",
-        name: "tagChoices",
-        message: "Select attributes to be added as tags, like key@value",
-        choices: choiceList
-    },
-    {
-        type: "checkbox",
-        name: "idChoice",
-        message:
-            "Select attributes to be used as the GeoJSON Feature ID (must be unique)",
-        choices: choiceList
-    }
-];
-
-export function createQuestionsList(object: any) {
-    for (let i = 0; i < 3 && i < object.features.length; i++) {
-        let j = 0;
-        for (let key in object.features[0].properties) {
-            if (i === 0) {
-                const desc =
-                    "" +
-                    (1 + j++) +
-                    " : " +
-                    key +
-                    " : " +
-                    object.features[i].properties[key];
-                choiceList.push({ name: desc, value: key });
-            } else {
-                choiceList[j].name =
-                    choiceList[j].name + " , " + object.features[i].properties[key];
-                j++;
-            }
-        }
-    }
-    return questions;
-}
-
-export function uploadData(
-    id: string,
-    options: any,
-    tags: any,
-    object: any,
-    isFile: boolean,
-    tagProperties: any,
-    fileName: string | null,
-    uid: string,
-    printFailed: boolean = false
-): any {
-    return new Promise((resolve, reject) => {
-        let upresult: any = { success: 0, failed: 0, entries: [] };
-        if (object.type == "Feature") {
-            object = { features: [object], type: "FeatureCollection" };
-        }
-
-        if (options.errors) {
-            printFailed = true;
-        }
-
-        if (options.assign) {
-            //console.log("assign mode on");
-            const questions = createQuestionsList(object);
-            inquirer.prompt(questions).then((answers: any) => {
-                if (options.ptag === undefined) {
-                    options.ptag = "";
-                }
-                options.ptag = options.ptag + answers.tagChoices;
-                if (options.id === undefined) {
-                    options.id = "";
-                }
-                options.id = options.id + answers.idChoice;
-                //console.log(options.ptag);
-                //console.log("unique key - " + options.id);
-                //Need to be inside if, else this will be executed before user choice is inserted as its async
-                uploadDataToSpaceWithTags(
-                    id,
-                    options,
-                    tags,
-                    object,
-                    false,
-                    options.ptag,
-                    fileName,
-                    options.id,
-                    upresult,
-                    printFailed
-                ).then(x => resolve(x)).catch((error) => reject(error));
-
-            });
-        } else {
-            uploadDataToSpaceWithTags(
-                id,
-                options,
-                tags,
-                object,
-                false,
-                options.ptag,
-                fileName,
-                options.id,
-                upresult,
-                printFailed
-            ).then(x => resolve(x)).catch((error) => reject(error));
-        }
-
-    });
-
-}
-
