+++ tsconfig.json
@@ -8,8 +8,7 @@
     "paths": {
       "@turf/voronoi": ["./declarations/@turf/voronoi.d.ts"],//voronoi typescript issue - https://github.com/Turfjs/turf/issues/1422
       "@turf/helpers": ["./declarations/@turf/helpers.d.ts"]
-    },
-    "sourceMap": true
+    }
     //"traceResolution": true
   }
 }
